Flyttet helper-metodene og valideringen tilbake inn i Controlleren, var usikker på hvor godt det funka..
